refactor(admin home): added footer and header to adminHome
