+++ app/src/routes.js
@@ -6,7 +6,7 @@ import Vehicles from './views/User/Vehicles/Vehicles.vue';
 import Login from './views/Login/Login.vue';
 import Register from './views/Register/Register.vue';
 import Reservation from './views/User/Reservation/Reservation.vue';
-import AdminHome from './views/Admin/Home.vue'
+import AdminHome from './views/Admin/Home/Home.vue';
 
 const routes = [
   {
