--- app/src/views/Admin/Home.vue
@@ -1,27 +1,0 @@
-<script setup>
-  import { CarOutlined, FileOutlined, UserOutlined } from "@ant-design/icons-vue"
-</script>
-
-<template>
-
-  <div class="admin-home-container" >
-      <a class="menu-item">
-        <router-link to="/admin/vehicles">
-          <CarOutlined/> Veiculos
-        </router-link> 
-      </a>
-      <a class="menu-item">
-        <router-link to="/admin/reservation">
-          <FileOutlined/> Reservas
-        </router-link>  
-      </a>
-      <a class="menu-item">
-        <router-link to="/admin/user"> 
-          <UserOutlined/> Usuários
-        </router-link>
-      </a>
-  </div>
-  
-</template>
-
-<style src="./Home.scss" lang="scss" scoped></style>
