+++ app/src/views/Admin/Home/Home.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors.scss";
+@import "../../../styles/colors.scss";
 
 .admin-home-container {
   display: grid;
